# openapi/registry-v1.yaml
openapi: 3.1.0
info:
  title: VVMP Registry API
  version: 0.1.0-draft
  description: |
    Draft machine-readable contract for VVMP registry implementations.
    This file is intended to be stack-neutral and usable by Next.js, FastAPI,
    Express, Go, Spring, ASP.NET, and other server implementations.
servers:
  - url: https://example.vvmp.org
paths:
...
  /api/v1/manifests/{manifestId}:
    get:
      summary: Get manifest by ID
      operationId: getManifestById
      parameters:
        - $ref: "#/components/parameters/ManifestId"
        - name: view
          in: query
          required: false
          schema:
            type: string
            enum: [public, full]
        - name: version
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Manifest response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManifestEnvelope"
        "404":
          $ref: "#/components/responses/Problem"
...
components:
  parameters:
    ManifestId:
      name: manifestId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Problem:
      description: Problem response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    ManifestEnvelope:
      type: object
      required: [manifest_id, manifest]
      properties:
        manifest_id:
          type: string
        version_id:
          type: string
        manifest:
          type: object
...
    Problem:
      type: object
      required: [title, status]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        instance:
          type: string